Fix bug related to error code 73
The bug is caused due to using the same matrix in multiple spots in the
BlockMat. Setting dirichlet boundary conditions would then change
multiple instances of matrices which is not correct, for example, for
matrices on the diagonal compared to the off-diagonal. This was fixed by
using copies everywhere. The last KSP modal eigenvalue problem doesn't
work because the matrix that requires inversion is singular.