Fix bug in initializing transient simulation from equilibrium
The bug occured because the impulse in psub was applied at time
zero which has no effect on the dynamics. The impulse has to be applied
at the second time for it to have an effect. I think this made the
simulations just simulate no motion, which caused the newton solver to
never reach the target residual.