Add explicit zeros to zero diagonal blocks of Hopf system
A lack of explicit zeros on the diagonal of the Hopf system was causing
the PETSc error 73. I think the error was being through when PETSc tried
to compute the LU decomposition but encountered a zero on the diagonal.
Since PETSc's LU decomposition doesn't pivot this probably lead to a
failure.